requirement.txt
appdirs==1.4.4
asgiref==3.2.7
audioread==2.1.8
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
Django==3.0.6
django-appconf==1.0.4
django-ckeditor==5.9.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==1.3.0
eyeD3==0.8.12
future==0.18.2
gym==0.17.2
hmmlearn==0.2.3
idna==2.10
imutils==0.5.3
joblib==0.16.0
jsonfield==3.1.0
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.33.0
lxml==4.5.2
matplotlib==3.1.2
mysqlclient @ file:///E:/pyCharmProject/venv/Lib/mysqlclient-1.4.6-cp38-cp38-win32.whl
numba==0.50.1
numpy==1.18.1
opencv-python==4.3.0.36
packaging==20.4
pandas==1.0.5
pandas-datareader==0.9.0
pilkit==2.0
Pillow==7.1.2
pooch==1.1.1
# Editable install with no version control (pyAudioAnalysis==0.3.2)
-e e:\ai\pyaudioanalysis-master
pycparser==2.20
pydub==0.23.1
pygame==1.9.6
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
requests==2.24.0
resampy==0.2.2
scikit-learn==0.23.2
scikit-surprise==1.1.0
scipy==1.5.1
simplejson==3.16.0
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
speech==0.5.2
sqlparse==0.3.1
surprise==0.1
swapper==1.1.2.post1
threadpoolctl==2.1.0
tk==0.1.0
tqdm==4.44.1
trim==0.3
urllib3==1.25.9
